feat: PoC dynamic typehash (#34)
* feat: PoC dynamic typehash

One problem with the witness field is that it is opaque to the user in
the EIP-712 typed data when they sign in their wallet. This commit adds
dynamic typehashes so integrators can provide the type data for the
witness field. Permit2 then uses this in the signature generation so
that the witness data can be parsed

* fix: change name of custom permit typehash

* fix: typo

* feat: separate witness name from type

* feat: remove vm parameter from permit test helper

* feat: remove witness from the other function

This commit removes witness from the normal transferFrom function so
that if provided it must have clear and explicit typing. Removes some
nice code deduplication but we can come back and clean it up later
i think

* fix: tests